Make AbstractTextRow more abstract
Nothing in AbstractTextRow implementation suggests it could only apply
to text based fields : I suggest renaming it and moving it to its own
file for readability.
